Add coercion prop option when using Vue > 1.0.12
Allow simpler notation for Boolean attributes in markup. Instead of
(in Jade):

```jade
    alert(v-bind:dismissable="true") message
```

You can now just type:

```jade
    alert(dismissable="true") message
```

If the value of the property is not "true", "false", "null" or "undefined"
then it will simply be passed as-is so that Vue can throw a warning.

If the version of Vue is older than 1.0.12 then the coercion simply won't
happen and you will have to specify props using the longer notation.