Correction to case insensitive comparisons in typeahead
Currently the value being compared is forced to lowercase, but the query string against which it is being compared isn't. this means that matches are possible only to lowercase input (this can be seen currently on the US states example on the documentation site). This fix forces both value and query string to lowercase (for case insensitive comparisons only) allowing upper, lower and mixed case searches to match when appropriate.